Wrote some rich interfaces to the bare-bones data model interfaces.
Given these rich interfaces, it also made sense for me to go ahead
and have createNode take an additional argument to specify the id
of a node, and then to have the rich interface generate one
automatically using UUIDs if desired.  The ability to specify the
id of a node in the tree will make lots of functionality much
easier to implement in a way that does not require extending the
basic data model interface itself.