Made the TreeController synchronous.
The more that I thought about it, the more that this
simply made the most sense from a user interface
perspective --- that is, it is better to do
as much as possible to keep the user interface
updated with the result of the most recent user
action then to return control to the user as soon
as possible after he makes a request to perform an
action.  Asynchronous actions should be treated as
a special case to be handled specially rather than
as the norm.  This also has the advantage of
removing some overhead so that small actions will
be completed faster and thus hopefully be much
less likely to result in any noticeable lag at all.